requirements.txt
pandas>=1.5.0
requests>=2.28.0
Flask>=2.3.0
python-frontmatter>=1.0.0
markdown>=3.4.0